Master Upload
This upload includes both the REST API and the windows phone 8.1 app
which uses the API.
This application allows the user to keep a personalised to do list.
First the user must use facebooks single sign on function to sign into
the app. They are then able to add to their list. This is done through
the REST API which updates a database hosted on the same server.
All web services were hosted on Microsoft Azure.